MySQL Bug #74286, Bug #19783949 - UPGRADING MYSQL FOR EXCEL FAILS FOR ALL VERSIONS SINCE 1.2.1
- Added an additional validation in the PerformActionOnRegistryView method for the Remove action to check if the add-ins registry key exists before attempting to delete it. (ExcelCustomAction\CustomAction.cs)
- Updated the customa action DLL on the MySQL for Excel project. (Installer\ExcelCustomAction.CA.dll)
- Changed the order and conditions of how custom actions are executed. The "Remove" actions are now executed when removing or upgrading the product and before the "Install" ones, which are executed on both new installations and upgrades. This fixes the problem reported in this bug where registry keys from previous installations were being left as is during an upgrade. (Installer\main.wxs)
- Added to the repository 2 files needed and that were missed in the bazaar > GIT migration. (ExcelCustomAction\Properties\Resources.Designer.cs, ExcelCustomAction\Properties\Resources.resx)

+++ ExcelCustomAction/CustomAction.cs
@@ -15,6 +15,8 @@
 // Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
 // 02110-1301  USA
 
+using System.Linq;
+
 namespace ExcelCustomAction
 {
   using System;
@@ -199,7 +201,7 @@ namespace ExcelCustomAction
               if (excelKey != null)
               {
                 var addinsKey = excelKey.OpenSubKey("Addins", true);
-                if (addinsKey != null)
+                if (addinsKey != null && addinsKey.SubKeyCount > 0 && addinsKey.GetSubKeyNames().Contains("MySQL.ForExcel"))
                 {
                   addinsKey.DeleteSubKeyTree("MySQL.ForExcel");
                   if (addinsKey.SubKeyCount == 0 && addinsKey.ValueCount == 0)
